solved problem with crash by reason of use BD in main treed, changed resource names in layouts.

+++ app/src/main/java/com/bereg/vacancyviewerapp/ui/activity/MainActivity.java
@@ -51,7 +51,7 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
                     return VacancyListFragment.getInstance();
                 case Screens.DETAILED_VACANCY_SCREEN:
                     Log.e(TAG, "DETAILED_VACANCY_SCREEN");
-                    return DetailedVacancyFragment.getInstance(data);
+                    return DetailedVacancyFragment.getInstance((Long) data);
             }
             Log.e(TAG, "NULL");
             return null;
